11_Mid_term/server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.4
langchain>=0.3.15
langchain-community>=0.3.15
langchain-openai>=0.3.2
langchain-qdrant>=0.1.1
openai>=1.58.1
langgraph>=0.2.67
python-dotenv==1.0.1
gtts==2.5.1
pyyaml==6.0.1
tiktoken>=0.6.0
qdrant-client>=1.7.0
sentence-transformers>=3.4.1 